debug: i*1000000 will exceed the numerical range of int

+++ tests/real_file_test.cpp
@@ -28,7 +28,7 @@ int main() {
       float z = i * 1.5;
 
       gyro_values.onsensor_timestamp_us = get_current_timestamp_us();
-      gyro_values.timestamp_ns = i * 1000000; // 模拟纳秒时间戳
+      gyro_values.timestamp_ns = (uint64_t)i * 1000000; // 模拟纳秒时间戳
       gyro_values.type =
           DumpSensorType::DUMP_SENSOR_TYPE_GYROSCOPE_UNCALIBRATED; // 示例类型
       gyro_values.data[0] = x;
@@ -40,7 +40,7 @@ int main() {
 
       RawImuDataDumpStruct accel_values;
       accel_values.onsensor_timestamp_us = get_current_timestamp_us();
-      accel_values.timestamp_ns = i * 1000000; // 模拟纳秒时间戳
+      accel_values.timestamp_ns = (uint64_t)i * 1000000; // 模拟纳秒时间戳
       accel_values.type = DumpSensorType::
           DUMP_SENSOR_TYPE_ACCELEROMETER_UNCALIBRATED; // 示例类型
       accel_values.data[0] = x * 2;
